requirement.txt
1.pytorch 1.7
2.transformers 4.6
3.onnxruntime